+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Patch notes
 
-Current `extended-click-outside` version: **0.0.2**
+Current `extended-click-outside` version: **1.0.0**
 
 ***
 
@@ -11,3 +11,15 @@ Current `extended-click-outside` version: **0.0.2**
 ### 0.0.2 (2022-02-05)
 
 * Adding .github/workflows directory to automate deployment.
+
+### 1.0.0 (2022-02-05)
+
+* Adding "@babel/plugin-transform-arrow-functions" dev-dependency.
+* Adding "@rollup/plugin-babel" dev-dependency.
+* Adding "babel-plugin-transform-imports" dev-dependency.
+* Adding "rollup" dev-dependency.
+* Adding "rollup-plugin-dts" dev-dependency
+* Adding rollup.config.mjs file as build configuration.
+* Adding some new keywords in "package.json" file.
+* Adding "dist" directory for current builds of package.
+* Сhanging the location of types in "package.json" file.